:committee-types configuration: https://github.com/metanorma/metanorma-mbxif/issues/1
