+++ lib/metanorma/standoc/front_committee.rb
@@ -6,14 +6,16 @@ module Metanorma
       end
 
       def committee_contributors(node, xml, agency, opts)
-        t = metadata_committee_types(node)
+        t = if c = node.attr("committee-types")
+              c.split(",")&.map(&:strip)
+            else metadata_committee_types(node)
+            end
         v = t.first
-        if committee_number_or_name?(node, v, "")
-          node.attr(v) or node.set_attr(v, "")
-          o = committee_contrib_org_prep(node, v, agency, opts)
-          o[:groups] = t
-          org_contributor(node, xml, o)
-        end
+        committee_number_or_name?(node, v, "") or return
+        node.attr(v) or node.set_attr(v, "")
+        o = committee_contrib_org_prep(node, v, agency, opts)
+        o[:groups] = t
+        org_contributor(node, xml, o)
       end
 
       def metadata_committee_types(_node)
@@ -42,11 +44,7 @@ module Metanorma
       end
 
       def contrib_committee_build(xml, agency, committee)
-        if name = org_abbrev.invert[agency]
-          committee[:agency_abbrev] = agency
-          agency = name
-        end
-        add_noko_elem(xml, "name", agency)
+        contrib_committee_build_agency(xml, agency, committee)
         s = committee
         loop do
           contrib_committee_subdiv(xml, s)
@@ -54,6 +52,14 @@ module Metanorma
         end
         add_noko_elem(xml, "abbreviation", committee[:agency_abbrev])
         full_committee_id(xml.parent)
+      end
+
+      def contrib_committee_build_agency(xml, agency, committee)
+        if name = org_abbrev.invert[agency]
+          committee[:agency_abbrev] = agency
+          agency = name
+        end
+        add_noko_elem(xml, "name", agency)
       end
 
       def contrib_committee_subdiv(xml, committee)
